Fix wildcard matching in people search
The wildcard for the first name was at the beginning so it would only match the end of a first name, which was weird.

e.g. `Dagny Aberkirder` was matching a search for `y aberkirder` but not `d aberkirder` - which is weird and counterintuitive

Move the wildcard to the end so partial matches always match from the beginning of the name.