Pull out separate `whereBillable` query method
This allows upstream services to do easier queries because they don't need to include all of the selects and joins in order to get the simple number, and can leave those off if not retrieving actual data rows.

Also does some additional filtering of returned fields in order to reduce the amount of extraneious data that is manually removed API side.