Add a fast prefilter to PIL transfer query
Running the sort/limit query on PIL transfer is an expensive operation to do before getting an establishment to compare because it gets done many times inside quite a large loop.

Check that the PIL has _ever_ been transferred first before doing repeated sort operations on the table.

This reduces execution time on the query by ~80%