Add migration to add RA to training and education PPLs
First migration sets the compulsory flag on all versions which have `training-licence` set to true.

Second migration re-calculates the RA date on the top-level project for any licences which may have been affected by the first migration - and is an almsot complete clone of the 20200710 migration to add RA dates to projects before, except it is limited to only those which are not _already_ set as having RA dates since this change will not unset any RA flags.