Make seeds non-random
When selecting licence-holders and named people ensure that the seeds are reproducible and so will produce a consistent state on every run.

Use the place name and project titles to create initial state when selecting an arbitrary role holder of licence holder. This means that the assignments will only change if the name or title is changed.