Add document cache model
Allows possibly slow functions to be cached in the database.

If a record has been generated in the 60 seconds then use that and do nothing else, otherwise try the query, and if it does not complete within 2 seconds then return the cached version, but allow the background function to continue and refresh the cache accordingly.