Include projects with additional availability in `profile.projects`
When calling `Profile.get()` in an establishment scoped query, include any projects in the related graph which have additional availability at the scoped establishment.

In order to do this, refactor the Project model to expose some additional query methods on the query builder class to support availability queries from outside the model - specifically in the `filter` property on the Profile=>Project relation.

Add a whole bunch of unit tests for all of this.