Allow changelog.changed_by to be nullable
In certain error scenarios the resolver tries to log an error into the changelog table, but without a user. This causes a second error to be thrown inside the error handler as the record cannot be inserted.

Allow the column to be null so the error record can be inserted with a null `changed_by` property.