Fix ROPs due date calculation query
Returns for 2021 on projects which have expired or been revoked in early 2022 were showing a due date of expiry/revocation + 28 days (i.e. some time in February) despite the fact that the correct due date for 2021 returns is 31st january because the projects were still active at the end of the year.

Update the query so that expired and revoked projects only have a due date of "end date + 28 days" if they ended before the end of the ROPs year.