Don't include deleted roles in place join
When building the search index for places it is preserving the role assignments even once the person has been removed from the role. In theory the place assignments should be deleted at the point of removing the role, but there's some lingering ones from before the code change that are shwoing up weirdly. Filter those out in the query.